creating obj
In Python, objects are created by instantiating a class. This means calling the class as if it were a function, which triggers the __init__ method to initialize the object's attributes. When an object is created, it gets a unique memory location, and its properties can be accessed or modified using dot notation. Multiple objects can be created from the same class, each with its own independent attributes. For example, if we define a Car class with attributes like brand and model, we can create different car objects such as car1 = Car("Toyota", "Corolla") and car2 = Car("Honda", "Civic").